telegram-bots/owner-bot/requirements.txt
python-telegram-bot>=21.0
python-dotenv==1.0.0
supabase>=2.0
schedule==1.2.0
anthropic
websockets>=14.0
httpx<0.28
